fix executable path for output and data folder
