+++ src/main.cpp
@@ -20,36 +20,6 @@
 #include "Utilities.hpp"
 #include "ProductToString.hpp"
 
-#include <filesystem>  // For filesystem operations in C++17
-
-
-#if defined(_WIN32) || defined(_WIN64)
-    #include <windows.h>
-    #include <cstring>
-#else
-    #include <libgen.h>  // For dirname()
-    #include <cstring>
-    #include <cstdlib>
-#endif
-
-// Function to get the directory of the executable
-std::string getExecutableDirectory(const char* argv0) {
-#if defined(_WIN32) || defined(_WIN64)
-    // Windows: Use GetModuleFileNameA to get the executable path
-    char buffer[MAX_PATH];
-    GetModuleFileNameA(NULL, buffer, MAX_PATH);  // Get the path of the executable
-    std::string path(buffer);
-    size_t found = path.find_last_of("/\\");  // Find the last '/' or '\\'
-    return path.substr(0, found);  // Return the directory part
-#else
-    // Linux/macOS: Use dirname from <libgen.h> to get the executable directory
-    char* path = strdup(argv0);  // Make a copy of argv[0] because dirname modifies the string
-    std::string dir = dirname(path);
-    free(path);
-    return dir;
-#endif
-}
-
 
 int main(int argc, char** argv) {
